feat: Create new function
Created a new function that calls one of the previous function depending on the user input. It takes 3 parameters- operand1, operand2 & operator - and then prompts the user to assign a value to each of the 3 variables. The value of operand1 & operand2 gets turned into an integer using the parseInt() function. And then a function is called depending on the value for operator.