feat: Actually make the calculator work
Started by creating an array to store the operands and a variable for the operator. And then created a function that sets the value for the operator depending on the button clicked. Pushes new elements to the operands array if any of the number buttons is clicked. Changed the logic in the operate function, if the length of the operands array is longer or equal to 2 and the operator returns a truthy value, it calls for the calculateResult function.The calculateResult function calls for one of the previously created math functions depending on the value of operator. The calculateResult function returns the value for result and which is turned to string and becomes the new displayValue. Updates the textContent and the operands array. If there are not enough operands or operators, it clears the display.