feat: Update the code
Turned the operator variable into an array named operators, which gets updated everytime an operator is input. Changed the logic in calculateResult again, now it has a loop which loops once for every item in the operators array, and the loop has if...else if conditionals that check for if the operators includes a certain operator, if it does it calls for a math function and removes the first element from the array, which is the current operator. The clear button now removes both the arrays along with the displayValue.