fix: Fix bug where calculations with multiple digited numbers created issues
Calculations with multiple digited numbets created issues. This occured because every time the user inputs a number, it was pushed into the array as a single element and got treated as a single digit number, even though it could be a part of a bigger number. But this was fixed by taking the pushing the inputs into an array, and turning all of them into a single digited or multiple-digited number using new function updateOperands() and pushing them into a different array everytime the user presses an operator button or the equals button