feat: Add backspace key and function and fix a bug
Fixed a bug where if the user presses the '=' button or the Enter key, a 0 is added to the operands array if the numInput array was empty. This was fixed by adding an if conditional in the function updateOperands() which returns the same operands array if numInputs is empty. Added a backspace button and keyboard input connected to handleBackspace() function, it uses if...else if conditionals and removes the last element from operators if the last character in displayValue is the same as the last element in the operator array by using .pop() method. For the operands and numInputs arrays, things are more complex. First the numInputs array is only triggered if it isn't empty, it creates a new array lastElementChars. And a new variable lastElement which is the last element in the numInputs array. This element is turned into a string and split into multiple characters and then assigned to a new variable characters and then it is pushed to lastElementChars. Now, if lastElementChars is not empty, it joins everything in the lastElementChars and converts it to number, and then assingns it to a new variable poppedNumber and then the last element from numInputs is removed using .pop() and the new poppedNumber is pushed into it. But if lastElementChars is empty, then the last element from numInputs is just removed using .pop(). The operands array follows the same mechanism but is only triggered if operands is not empty but numInputs is. And then the displayValue is updated by removing the last character using slice() method.