fix: Fix bug that occurs when trying to diplay '.' and round result if its not an integer
Fixed a bug that occurs when the user inputs '.', the updateDisplay function was not working as expected. Fixed it by changing the updateDisplay function and creating a new function that calls the updateDisplay function and the setNumInput function if numInputs is empty. Attached this new function to eventlisteners and handleKeyPress function that deals with '.', also updated the operate function to round the result to 2 places if it is not an integer.